pyproject.toml
[project]
name = "bulk-metadata-scaler-app"
version = "0.1.0"
description = "Bulk metadata enrichment app for Atlan - Updates asset metadata based on reference files"
authors = [
    {name = "Atlan Team"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "atlan-application-sdk[workflows,pandas]>=1.0.0",
    "openpyxl>=3.1.0",
    "python-multipart>=0.0.9",  # For file uploads in FastAPI
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bulk_metadata_scaler_app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

